requirements.txt
sounddevice
pyaudio
python-levenshtein
vosk
pvporcupine
numpy
pygame

# python3 -m venv venv
# source venv/bin/activate.fish
# pip install -r requirements.txt
# python3 vosc.py